refactor : 이벤트 큐이름 변경에 따른 코드 수정

+++ src/main/java/com/example/surveyapi/domain/statistic/domain/statisticdocument/dto/SurveyMetadata.java
@@ -22,7 +22,7 @@ public class SurveyMetadata {
 		String questionType,
 		Map<Long, String> choiceMap
 	) {
-		public Optional<String> getChoiceText(Long choiceId) {
+		public Optional<String> getChoiceText(Integer choiceId) {
 			return Optional.ofNullable(choiceMap.get(choiceId));
 		}
 	}
